+++ test/app.ts
@@ -3,7 +3,7 @@
 // allow the TS compiler to successfully import
 // and interact with the JS client
 
-import { Client } from '..';
+import { Client } from './tmp/package';
 const client = new Client('API_KEYYY')
 client.notify(new Error('flop'))
 client.setUser('123', 'B. Nag', 'bugs.nag@bugsnag.com')